API-80: Extract armCyclesWithScheduleMarkers into service/daemon/markers.
